apr/pyproject.toml
[tool.poetry]
name = "APR"
version = "0.1.0"
description = "Automatic Personality Recognition service"
authors = ["Matías Apablaza <mapablaza@itba.edu.ar>", "Yash Mehta <yashsmehta95@gmail.com>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.9, <3.12"
boto3 = "^1.28.65"
pymongo = "^4.5.0"
python-dotenv = "^1.0.0"
openai = "^1.35.3"
tensorflow-cpu = "^2.13.0"
numpy = "^1.22"
scikit-learn = "^1.3.0"
pandas = "^2.0.3"
transformers = {extras = ["torch"], version = "^4.31.0"}
torch = {version = "^2.0.1+cu118", source = "pytorch-gpu-src"}
tweet-preprocessor = "^0.6.0"

[[tool.poetry.source]]
name = "pytorch-gpu-src"
url = "https://download.pytorch.org/whl/cu118"
priority = "explicit"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
